+++ src/Core/Dfe.Complete.Application/Notes/Interfaces/INoteWriteRepository.cs
@@ -6,6 +6,7 @@ namespace Dfe.Complete.Application.Notes.Interfaces;
 public interface INoteWriteRepository
 {
     Task<Note?> GetNoteByIdAsync(NoteId noteId, CancellationToken cancellationToken);
-    Task UpdateNoteAsync(Note note, CancellationToken cancellationToken);
+    Task CreateNoteAsync(Note note, CancellationToken cancellationToken);
     Task RemoveNoteAsync(Note note, CancellationToken cancellationToken);
+    Task UpdateNoteAsync(Note note, CancellationToken cancellationToken);
 }